Moved a test that needs BinaryFormatter to a different project
