+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/ClipboardTests.cs
@@ -460,20 +460,10 @@ public partial class ClipboardTests
     }
 
     [WinFormsFact]
-    public void ClipBoard_SetData_CustomFormat_Color()
-    {
-        using BinaryFormatterScope scope = new(enable: true);
-        string format = nameof(ClipBoard_SetData_CustomFormat_Color);
-        Clipboard.SetData(format, Color.Black);
-        Assert.True(Clipboard.ContainsData(format));
-        Assert.Equal(Color.Black, Clipboard.GetData(format));
-    }
-
-    [WinFormsFact]
     public void ClipBoard_SetData_CustomFormat_Color_BinaryFormatterDisabled_SerializesException()
     {
         using BinaryFormatterScope scope = new(enable: false);
-        string format = nameof(ClipBoard_SetData_CustomFormat_Color);
+        string format = nameof(ClipBoard_SetData_CustomFormat_Color_BinaryFormatterDisabled_SerializesException);
 
         // This will fail and NotSupportedException will be put on the Clipboard instead.
         Clipboard.SetData(format, Color.Black);
@@ -489,9 +479,8 @@ public partial class ClipboardTests
         {
             formatter.Serialize(stream, new object());
         }
-        catch (NotSupportedException ex)
+        catch (NotSupportedException)
         {
-            Assert.Equal(ex.Message, value.Message);
             return;
         }
